# swagger.yaml
openapi: 3.0.3
info:
  title: Ecommerce Project API
  version: 1.0.0
  description: |
    OpenAPI specification generated from Fastify route registrations and controller payload handling.

    Authentication model used by this API:
    - `x-user-id` header is required for authenticated routes.
    - `x-user-role: admin` is required for admin-only routes.
servers:
  - url: http://localhost:3000
    description: Local development
...
paths:
...
  /users/:
    get:
      tags: [Users]
      summary: List all users (admin)
      operationId: getAllUsers
      security:
        - UserIdHeader: []
          UserRoleHeader: []
      responses:
        '200':
          $ref: '#/components/responses/ApiSuccess'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags: [Users]
      summary: Create a user
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          $ref: '#/components/responses/ApiSuccess'
        '400':
          $ref: '#/components/responses/BadRequest'
  /users/{id}:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    get:
      tags: [Users]
      summary: Get user by ID
      operationId: getUserById
      security:
        - UserIdHeader: []
      responses:
        '200':
          $ref: '#/components/responses/ApiSuccess'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Users]
      summary: Update user
      operationId: updateUser
      security:
        - UserIdHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          $ref: '#/components/responses/ApiSuccess'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Users]
      summary: Delete user (admin)
      operationId: deleteUser
      security:
        - UserIdHeader: []
          UserRoleHeader: []
      responses:
        '200':
          $ref: '#/components/responses/ApiSuccess'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    UserIdHeader:
      type: apiKey
      in: header
      name: x-user-id
      description: Positive integer user ID
    UserRoleHeader:
      type: apiKey
      in: header
      name: x-user-role
      description: Use value `admin` for admin-only endpoints
  parameters:
    IdPath:
      in: path
      name: id
      required: true
      schema:
        type: integer
  responses:
    ApiSuccess:
      description: Successful operation
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: true
              data:
                nullable: true
              message:
                type: string
...
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
        error:
          type: string
    CreateUserRequest:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        role:
          type: string
          enum: [customer, admin]
        avatar:
          type: string
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        role:
          type: string
          enum: [customer, admin]
        avatar:
          type: string
...